+++ HISTORY.rst
@@ -11,6 +11,7 @@ Upcoming
 - Travis CI testing for CPython 3.5 and PyPy 2.6.0.
 - Added a logging network class that logs requests and responses.
 - Added new options for auth classes, including storing tokens in Redis and storing them on a remote server.
+- Stream uploads of files from disk.
 
 1.2.2 (2015-07-22)
 ++++++++++++++++++